Move unit test files into unit test folder
